code/Biofeedback/requirements.txt
pyserial==3.5
python_osc==1.8.3
